Refactor 'views.py' into Multiple Files
